Kiln_lib: Split library into separate files and fixed some lints
